Use finalize on julia 0.4 to make memory-management much more robust
On julia 0.3 it seems basically impossible to pass this test:

    for i = 1:100
        print(i, "..")
        include("runtests.jl")
    end

because both the julia address and the pointer address may on
occasion be the same; if you allocate a Ptr, do a device_reset,
and there's a new allocation that duplicates both pointer addresses,
the first object's finalizer may run and tromp on the new object.

finalize fixes that. Only works on julia 0.4, though.